7-8/requirements.txt
app~=0.0.1
Jinja2~=3.1.2
pandas~=1.5.1
asyncpg~=0.27.0
python-magic~=0.4.27